// packages/companion-module-base/src/host-api/api.ts
export interface FeedbackInstanceBase {
	id: string

	// If this is pending being run through upgrade scripts, the version it needs upgraded from is tracked here
	upgradeIndex: number | null
	disabled: boolean

	feedbackId: string // aka 'type'
	options: OptionsObject
}
// ...
export interface ActionInstanceBase {
	id: string

	// If this is pending being run through upgrade scripts, the version it needs upgraded from is tracked here
	upgradeIndex: number | null
	disabled: boolean

	actionId: string // aka 'type'
	options: OptionsObject
}
export interface ActionInstance extends ActionInstanceBase {
	controlId: string
}

export interface UpdateActionInstancesMessage {
	actions: { [id: string]: ActionInstance | null | undefined }
}

export interface UpgradeActionInstance extends Omit<ActionInstanceBase, 'options'> {
	options: OptionsObject

	controlId: string
}
export interface UpgradeFeedbackInstance extends Omit<FeedbackInstanceBase, 'options'> {
	options: OptionsObject

	isInverted: boolean

	/**
	 * Only used as an output from the module, when the feedback is being converted to a boolean feedback
	 */
	style?: Partial<CompanionFeedbackButtonStyleResult>

	controlId: string
}

export interface UpgradeActionAndFeedbackInstancesMessage {
	actions: UpgradeActionInstance[]
	feedbacks: UpgradeFeedbackInstance[]
	defaultUpgradeIndex: number | null
}

export interface UpgradeActionAndFeedbackInstancesResponse {
	updatedConfig: unknown
	updatedSecrets: unknown
	updatedActions: UpgradeActionInstance[]
	updatedFeedbacks: UpgradeFeedbackInstance[]
	latestUpgradeIndex: number
}